refactor: move migrations exclusive functions into the migration
