+++ virtool/analyses/utils.py
@@ -35,26 +35,6 @@ async def attach_analysis_files(
     return {**document, "files": [result.to_dict() for result in results]}
 
 
-def check_nuvs_file_type(file_name: str) -> str:
-    """
-    Get the NuVs analysis file type based on the extension of given `file_name`
-
-    :param file_name: NuVs analysis file name
-    :return: file type
-
-    """
-    if file_name.endswith(".tsv"):
-        return "tsv"
-
-    if file_name.endswith(".fa"):
-        return "fasta"
-
-    if file_name.endswith(".fq"):
-        return "fastq"
-
-    raise ValueError("Filename has unrecognized extension")
-
-
 def find_nuvs_sequence_by_index(
     document: Dict[str, Any], sequence_index: int
 ) -> Optional[str]:
